Organized location related models into dedicated files and fixed a bug caused by hiding progress bar in the finally block.
