+++ app/data/location/LocationGroup.ts
@@ -1,0 +1,4 @@
+export default interface LocationGroup {
+  id: string
+  name: string
+}
