+++ app/data/location/LocationType.ts
@@ -1,0 +1,6 @@
+import {LocationTypeCode} from "./LocationTypeCode";
+
+export default interface LocationType {
+  description?: string
+  locationTypeCode: LocationTypeCode
+}
